Alright, couldn't fix Travis so need to abstract Files API to mock it.
The java NIO APIs are unfortunately made to be static, so mocking often
does not work. So we have to go ahead and create our own interfaces
and implementations so we can mock them.

This now allows us to fully test the ImageRepository deletionl logic without
actually writing anything on the filesystem.