Fix TV shutdown then re-connect issue.
A socket does not know when the connection is closed when the other end simply
disappears. This happens when the TV is shutdown. In this case the bugs was easy
to reproduce: Shut down the TV through winston, restart the TV and then try to
shut it down again: It wouldn't work because the socket would not realize it was
no longer connected. This fix will recreate the socket if the remote wants it to
reconnect, thus solving the issue.